cjson: bug fix - use -fPIC instead of -fpic
This change is to fix the autobuild failure at:
http://autobuild.buildroot.net/results/45e2c08d5e3c868b2d7fdf26a7c5f88de5ff8f61/build-end.log
I can't actually reproduce this bug, because my machine cannot
run the 64-bit microblaze toolchain.  However, the log makes
it clear that -fPIC is needed, and grepping for fPIC and fpic
under packages makes it clear that cjson should have used -fPIC
anyway.  So even if the bug isn't fixed, it must surely be improved.

Signed-off-by: Danomi Manchego <danomimanchego123@gmail.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>