lcdapi: fix autobuilder failures
Version 0.3 of the lcdapi package does not correctly handle CXXFLAGS passed
in from the environment/command line. In particular, the passing in of CXXFLAGS
clobbers the -fPIC flag which is set in the top-level lcdapi makefile.

The following autobuild logs highlight the problem.

  http://autobuild.buildroot.net/results/9a47b4a70d29308f668d65d983ba6146c0c56365/build-end.log

  http://autobuild.buildroot.net/results/4a4f6e2e7f80d5a23c9d26069094e9dafb1468e7/build-end.log

  http://autobuild.buildroot.net/results/0d96c207a9db44cf2791a66cad5162a739e7053a/build-end.log

This has been fixed upstream in version 0.4; this patch simply bumps the
version number in the Buildroot lcdapi package.

Signed-off-by: Simon Dawson <spdawson@gmail.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>