libtirpc: add patches to fix a number of issues
libtirpc-0005-rpcent-mark-getrpcbyname-name-argument-as-const-char.patch
fixes build failures like
http://autobuild.buildroot.org/results/7a3a751fe02c639ca75c575ca7fe20a72372b8d6/build-end.log.

libtirpc-0006-rpcent-remove-prototypes-of-reentrant-variants.patch
ensures that functions not implemented by libtirpc are not described
in a header file. It also allows to remove those prototypes that were
not matching the functions available in glibc or uClibc, causing
mismatch in prototypes.

libtirpc-0007-doc-Makefile.am-fix-out-of-tree-installation.patch is a
minor fix.

The following patches (8 to 9) allow libtirpc to provide sufficient
things to be able to build rpcbind on top of it.

All these patches have been submitted upstream on the libtirpc-devel@
mailing list.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Acked-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>