package/dvb-apps: fix downloading
As we're using the sources from the Mercurial repository, and Buildroot
does support retrieving from a Mercurial repository, there is no need
to try (and fail1) getting the archive from the Mercurial built-in
tarball mechanism.

(Note: I was beaten by this because I had a cached copy locally, left
after the previous tvheadend-vampirises-files-from-toher-packages attempt,
that I forgot to delete before testing. Ouch...)

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>