xapp_xinit depends on mcookie, not xserver_xorg-server
mcookie is not actually used by the Xorg server; it is used by xinit to
generate a xauth cookie before starting the server.

Verified with 'ack-grep -a mcookie build/x*' in an output directory with all
packages selected. The only other package that matches is x11vnc: it does a
runtime check for mcookie and falls back on /dev/random if it's missing.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>