urg: fix license
The urg.mk file mentions a LGPLv3+ license. However, none of the
source code in the urg package carries a copyright notice that
indicates the "v3 or later". The only license information we have is
the COPYING file, that contains the text of the LGPLv3
license. Therefore, we should probably assume that the code is under
LGPLv3 only.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>