libgtk2: use non-host specific patch
The libgtk2 package was using a host-<foo>.patch that should be
applied only on the host variant of the package. Unfortunately, with
the patch model rework, this doesn't work anymore: Buildroot tries to
apply the patch twice, for some reason.

But instead of fixing the patch model, it is probably a lot better to
fix this patch itself, which is the only usage of host-only patch in
Buildroot.

So instead of simply removing code that detects dependencies in gtk2's
configure.in script, we use a condition based on the value of
gdktarget. And it turns out that it makes the patch shorter.

However, it requires autoreconfiguring the libgtk2 target package,
because we are now modifying its configure.in.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>