rt-tests: mention that NPTL thread implementation is needed
This patch superseds http://patchwork.ozlabs.org/patch/150161/, but
does it in a different way. Since we don't have a way of selecting the
thread implementation for external toolchains, and it sounds a bit too
heavy to add new config options just for the sake of rt-tests, we
instead simply add some help text to the rt-tests package explaining
that NPTL thread implementation is needed.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>