dnsmasq: add missing dependency on host-gettext
When idn support is enabled, the dnsmasq build procedure requires
xgettext, which is provided by host-gettext.

Moreover, dnsmasq.mk was not passing $(TARGET_MAKE_ENV) in the
environment, so $(HOST_DIR)/usr/bin was not in the PATH. For that
reason, the dnsmasq Makefile could not find the xgettext program.

Fixes:
http://autobuild.buildroot.org/results/f7feaa818500a369a6c4c5fd8899c168f21b1460/build-end.log.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>