rpi-firmware: Fixup config.txt
The default name of the kernel binary for the RaspberryPi is kernel.img,
it's only a zImage kernel renamed.  Since Buildroot will build a zImage, the
default config should use the zImage name instead of forcing the user to
rename a file.

This patch also remove the cmdline argument rootfstype. There is no point
forcing the user to use ext4 for its root partition.

Signed-off-by: Maxime Hadjinlian <maxime.hadjinlian@gmail.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>