Unhide lttng-tools and lttng-modules
When using lttng-tools for userland tracing with lttng-libust lttng-modules
is not required, thus a dependency on building lttng-modules and a kernel is
overkill for lttng-tools.  It also hides it from a user not wanting to build
a kernel.  A comment has been added to lttng-modules to show a user that
lttng-modules is dependent on a kernel build.

Signed-off-by: Olivier Schonken <olivier.schonken@gmail.com>
Acked-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>