iptables: fix build after linux-headers changes
As reported by 'scanf' on IRC, the iptables now fails to build because
of a missing <asm/cache.h> header in the kernel headers. The package
directly uses $(LINUX_HEADERS_DIR), which is not correct: this is the
source directory of the kernel, not where they are installed, and in
addition, this directory is empty/does not exist when using external
toolchains.

Instead, we use $(STAGING_DIR)/usr, which contains the kernel headers,
as part of the toolchain headers.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>