package/weston: fix configure.ac to check for wayland-scanner
configure whines while checking for wayland-scanner.

wayland-scanner is used to generate the protocol parser C files from
the protocol definition XML files.

weston has a hard-dependency on wayland-scanner, so it can regenerate
its shell/mouse/keyboard/... "handlers".

Since we're using a tarball, those protocol files are already generated
and up-to-date, but the check is hard-coded and unconditional. If
wayland-scanner is missing, configure fails.

We could well patch away this check, but we'd have to carry and maintain
it probably for ever.

Better to fix it: add a patch from upstream weston to fix configure
whining.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>